feat: Add sticky sessions (#39)
Add support for localStorage to save session data (key piece of data here
is the id of the root replay event). Now sessions will expire after 15
minutes of inactivity.

Also simplifies the creation of `sentry-replay` transactions. There will
now only be 1 `sentry-replay` (root) transaction for a session. The
`replayId` is the id of the root transaction and is no longer different
than the rootReplayId.