fix: Fix resetting initial state when loading new session (#147)
This resets initial state when we load a new session (e.g. due to inactivity). I believe this is what is causing our durations to be all over the place. If a user starts a new session, it should reset the initial state, otherwise the replay start time will always be the whenever the tab was first opened, instead of when a new session is created.