feat: Remove sendBeacon() usage  (#152)
Theres not too much of a reason to use this, lets keep things simple and just use fetch which is well supported. I believe I added this to capture the window unload event, but even that does not work consistently anyway.

I was also seeing these warnings https://chromestatus.com/feature/5629709824032768 -- which is harmless in our case as it seems to block the *response* which we don't care about.