feat: Inspect events for earliest timestamp (#156)
We need to inspect events before they hit the event buffer (bc of compression) and keep track of the timestamp of the earliest event. This can happen because entries in the Performance Observer are buffered so they can occur before the integration is initialized.

This also fixes replay update events where we pass the replay start time, but we need to defined the timestamp, as that is used to calculate the finishedAt timestamp.