feat: Do not alter time.
Stop messing with time so much.

With the introduction of `replay_start_timestamp` and keeping track of the earliest timestamp, we can simplify `flushUpdate`, so that we are not attempting to preserve timestamps. `captureReplayUpdate`s timestamp should now always be after all of the breadcrumbs that have been collected thus far.