feat: Do not alter time. (#169)
Stop messing with time so much.

With the introduction of `replay_start_timestamp` and keeping track of the earliest timestamp, we can simplify `flushUpdate`, so that we are not attempting to preserve timestamps. At the time of flushing, we would capture memory metrics at that current timestamp, which was after the "preserved" timestamp. This caused our durations to be incorrect (which is calculated using the preserved timestamp as the end timestamp).

`captureReplayUpdate`s timestamp should now always be after all of the breadcrumbs that have been collected thus far. 


Before:
![image](https://user-images.githubusercontent.com/79684/187984122-1403f503-a8ba-4048-83e9-64a170d43fe4.png)

After:
![image](https://user-images.githubusercontent.com/79684/187984206-ae0d402a-e353-472b-b2f8-5fbc19f15685.png)