feat: Remove throttled flush in favor of only using debounced flush
A debounced flush was added in https://github.com/getsentry/sentry-replay/pull/210 -- remove the throttled flush call in favor of the debounced version. This will give a more consistent flushing behavior by using only one debounced flush. The throttle was also set at 1 second, which is a bit too close in time to the previous flush.