fix: Remove typescript `paths` aliases
We were using the tsc config `paths` for import aliases, *but* tsc does not rewrite paths. This needs to be handled in a build tool, in our case we could use a plugin like this https://www.npmjs.com/package/@rollup/plugin-alias. However, we are going to just remove the usage of import aliases altogether as there was mixed usage of it and I am opting to keep things a bit simpler for now.