fix: Treat sessions with `segmentId:0` as new session (#271)
We have lots of replays that are missing `replay_start_timestamp` on
segment 0. For some reason, some sessions with `segmentId:0` are not
being considered a "new" session. ~Ideally we would fix this but for
now~, add some debug logging and forcefully treat session as "new" if
after loading a session, it still has `segmentId:0`.

I've gone ahead and removed the new session flag, as it's not needed,
instead we can assume that segment 0 requires the start timestamp.

Fixes https://github.com/getsentry/sentry/issues/40984