feat: For error replays, reset state on checkout
Follow-up to https://github.com/getsentry/sentry-replay/pull/268#discussion_r1013504603

The session is always started immediately on pageload/init, but for
error-only replays, it should reflect the most recent checkout
when an error occurs.

Checkouts happen on an interval so, clear any state that happens before the current
checkout and update the session started time on every checkout.