feat: Fix initialState for error sessions  (#281)
There was an incorrect assumption that loading from session storage
implied that the session was new. We were only initializing initialState
in this "new" case, instead we should initialize whenever `start()` is
called.

Also fixes a bug where exceptions were being obscured because we were
always capturing a new/created error instead of capturing the error that
was thrown.