replays: increment segment id after finishing buffer (#306)
- since in some cases we will not be able to finish sending a replay
because we can't run `async` operations in some `visibilitychange`
situations, don't increment the segment_id until after we get the data
from event buffer. This should help in a bunch of cases where right now
we increment and miss the `0`th segment, and additionally help us track
down other problems with our flush/transport flow. in order to actually
get the data in this situation we'll have to re-architect our event
buffer and not do any async operations in our flush flow.